move all user classes to same file
